+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.12.0] - 2023-11-18
+
+### Added
+
+- Add German localization.
+
 ## [0.11.0] - 2023-11-18
 
 ### Added